+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 
 ### Добавлено
 - Класс `BaseProvider` для работы с внешними сервисами;
+- Классы тестов `TestBaseProvider`, `TestCheckResponseStatus` под `BaseProvider`;
 
 
 ## [0.4.4]